Fix typing

+++ vsdeband/funcs.py
@@ -80,7 +80,7 @@ def mdb_bilateral(
 
 def masked_deband(
     clip: vs.VideoNode, radius: int = 16,
-    thr: int | list[int] = 96, grain: float | list[float] = [0.23, 0],
+    thr: float | list[float] = 96, grain: float | list[float] = [0.23, 0],
     sigma: float = 1.25, rxsigma: list[int] = [50, 220, 300],
     pf_sigma: float | None = 1.25, brz: tuple[int, int] = (0.38, 0.68),
     rg_mode: RemoveGrainModeT = RemoveGrainMode.MINMAX_MEDIAN_OPP,
@@ -102,7 +102,7 @@ def masked_deband(
 
 def pfdeband(
     clip: vs.VideoNode, radius: int = 16,
-    thr: int | list[int] = 120, grain: float | list[float] = 0.0,
+    thr: float | list[float] = 120, grain: float | list[float] = 0.0,
     lthr: int | tuple[int, int] = (76, 0), elast: float = 2.5,
     bright_thr: int | None = None, scaler: ScalerT = Spline64,
     prefilter: Prefilter | VSFunction = partial(Prefilter.SCALEDBLUR, scale=1, radius=2),
